Client-N2_scraper/configuration/tank-scraper/requirements.txt
selenium
webdriver-manager
requests
pyyaml